341: break getAccelValues up into blocks in start.js
